+++ ui/pointer/vr3dscene/vr3dscene.gd
@@ -1,8 +1,6 @@
 extends Spatial
 
 func _ready() -> void:
-	SignalManager.connect("settings_changed", self, "_on_settings_changed")
-
 	adjust_position()
 	adjust_size()
 
@@ -10,7 +8,7 @@ func _physics_process(delta):
 	if $ARVROrigin/Controller/RayCast.is_colliding():
 		pass
 	
-func _on_settings_changed() -> void:
+func set_configuration(config : Dictionary, widgets : Array, render_target_size : Vector2) -> void:
 	adjust_position()
 	adjust_size()
 	
@@ -29,9 +27,9 @@ func adjust_position() -> void:
 	var rotation_z : float = SettingsManager.get_value("user", "overlay/rotation_z", DefaultSettings.get_default_setting("overlay/rotation_z"))
 	
 	var areatransform : Transform = Transform(Basis(Vector3(0, 0, 0)), Vector3(position_x, position_y, position_z))
-	areatransform.basis = areatransform.basis.rotated(Vector3(1, 0, 0), rotation_x + 1.5)	
-	areatransform.basis = areatransform.basis.rotated(Vector3(0, 1, 0), rotation_y)	
-	areatransform.basis = areatransform.basis.rotated(Vector3(0, 0, 1), rotation_z)	
+	areatransform.basis = areatransform.basis.rotated(Vector3(1, 0, 0), rotation_x + 1.5)
+	areatransform.basis = areatransform.basis.rotated(Vector3(0, 1, 0), rotation_y)
+	areatransform.basis = areatransform.basis.rotated(Vector3(0, 0, 1), rotation_z)
 
 	areatransform = areatransform.orthonormalized()
 	$ARVROrigin/OverlayArea.transform = areatransform