+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.co##m/en/1.0.0/
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED] - 2019-XX-XX
+### Added
+- Added styling for connections tab
+
 ### Changed
 - Update code to use the new multi-overlay implementation of godot_openvr
-- Update to Godot 3.2 beta
+- Update to Godot 3.2 rc3
 - Improvements to the README
 
 ### Fixed
@@ -16,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed "undim on stare"
 - Fixed not being able to accept login on enter
 - Fixed saving of "Play chime on event" checkbox
+- Fixed Twitch chat widget SSL mode crashing the overlay
+- Fixed overlay crashing when host is reveived on Twitch
 
 ## [0.0.3-alpha - 2019-09-15
 ### Added